Set CLSID of RootStorage to outlook clsid
(as far as I can tell) a CLSID is used an identifier for file types/applications/folders/whatever else, or as MSDN puts it: "A CLSID is a globally unique identifier that identifies a COM class object. If your server or container allows linking to its embedded objects, you need to register a CLSID for each supported class of objects." (real helpful...).

By setting the CLSID on the RootStorage to {0006-F03A-0000-0000-C000-0000-0000-0046} (sans punctuation), windows/outlook will interpret the msg file correctly and import it, rather than just interpreting it as a "file" and not actually parsing it. This behaviour *seems* consistent. There may be some deeper thing going on that should be taken care of...